修改交易详情节点名称获取

+++ manager/src/main/java/com/platon/browser/util/TxInfoResolver.java
@@ -15,7 +15,7 @@ public class TxInfoResolver {
     @Data
     private static class ResolveResult{
         String nodeId,nodeName;
-        Integer voteCount=0;
+        long voteCount=0;
         BigDecimal ticketPrice=BigDecimal.ZERO,deposit=BigDecimal.ZERO;
     }
     public static void resolve(String txType,String txInfo,String value,Object target){
